requirements.txt
# requirements.txt
# Core data processing & finance
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0

# Plotting & visualization
plotly>=5.15.0
kaleido>=0.2.1          # required for plotly static image export (fig.write_image)

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Kafka
kafka-python>=2.0.0

# Redis caching (optional but used in code)
redis>=5.0.0

# Email
secure-smtplib>=0.1.0   # optional, but helps with modern SMTP

# Utilities & concurrency
requests>=2.28.0
selenium>=4.10.0        # if you still use Grafana snapshot / web capture
webdriver-manager>=4.0.0  # auto-manages ChromeDriver for Selenium

# Prophet forecasting (if you re-add it later)
prophet>=1.1.0          # optional - uncomment if needed
pystan>=2.19.1.1        # prophet dependency (sometimes needed)

# Development / debugging (optional)
python-dotenv>=1.0.0    # for loading .env file
